refactor: wip: snapshot
