--- pipeline/checksums/doc.go
@@ -1,3 +1,0 @@
-// Package checksums provides a Pipe that creates .checksums files for
-// each artifact.
-package checksums
