+++ pipeline/piper.go
@@ -14,9 +14,9 @@ type Piper interface {
 	Run(ctx *context.Context) error
 }
 
-// ErrSkipPublish happens when skip publish is set and a pipe is refusing
-// to proceed because of that.
-var ErrSkipPublish = Skip("--skip-publish is set")
+// ErrSnapshotEnabled happens when goreleaser is running in snapshot mode.
+// It usually means that publishing and maybe some validations were skipped.
+var ErrSnapshotEnabled = Skip("disabled during snapshot mode")
 
 // IsSkip returns true if the error is an ErrSkip
 func IsSkip(err error) bool {
